{{cookiecutter.project_name_slug}}/pyproject.toml
[tool.stela]
default_environment = "development"
config_file_extension = "TOML"
config_file_prefix = "env-"

[tool.coverage.report]
fail_under = 80

[tool.pytest]
addopts = '--pydocstyle'

[tool.black]
target-version = ['{{ cookiecutter.black_target_version }}']
include = '\.pyi?$'
exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | \venv
    | \.aws-sam
    | _build
    | buck-out
    | build
    | dist
    | node_modules
    | dependencies
  )/
)
'''

[tool.poetry]
name = "{{cookiecutter.model_name_slug}}"
version = "1.0.0"
description = "{{ cookiecutter.project_description }}"
authors = ["Chris Maillefaud <chris@megalus.com.br>"]

[tool.poetry.dependencies]
python = "^{{ cookiecutter.python_version }}"
arrow = "*"
aws_lambda_context = "*"
loguru = "*"
marshmallow = "*"
marshmallow-pynamo-db = "*"
marshmallow_enum = "*"
pynamodb = "*"
pynamodb-attributes = "*"
sentry-sdk = "*"
stela = "*"
typing_extensions = "*"

[tool.poetry.dev-dependencies]
aws-sam-cli = {version="*", markers="platform_system != 'Windows'"}
boto3 = "*"
boto3-stubs = "*"
black = {version = "*", allow-prereleases = true}
factory_boy = "*"
freezegun = "*"
ipdb = "*"
pre-commit = "*"
pytest-black = "*"
pytest-cov = "*"
pytest-flake8 = "*"
pytest-freezegun = "*"
pytest-mock = "*"
pytest-mypy = "*"
pytest-pydocstyle = "*"
pytest-watch = "*"
pytest-xdist ="*"
wheel = "*"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
